feat: update app to use pg instead of sqlite
