+++ itu-minitwit-golang/.env.example
@@ -1,3 +1,9 @@
 PORT=8080
-DB_PATH=./data/myapp.db
 ENVIRONMENT=development
+
+
+DB_HOST=
+DB_USER=
+DB_PASSWORD=
+DB_NAME=
+DB_PORT=
